Create README.md
The Todo list project is a web-based application created using HTML, CSS, and JavaScript. The user interface of the application is designed using HTML and CSS, and it includes an animated background to enhance the visual appeal.

The main functionality of the application is to allow users to add and delete tasks, as well as mark them as complete by using a checkbox. JavaScript is used to handle these user interactions and to manipulate the content of the web page dynamically.

Users can add tasks by typing them into a text input field and clicking on an "Add" button. This action triggers a JavaScript function that adds the new task to a list on the page. Each task item includes a checkbox that can be checked to mark the task as complete. Users can also delete tasks by clicking on a "Delete" button that appears next to each task item when the user hovers over it.

Overall, this Todo list project provides a simple and user-friendly way for individuals to keep track of their tasks and manage their time effectively.